pyproject.toml
[project]
name = "user-state-probes"
version = "0.1.0"
description = "Linear probes for LLM internal representations of user-state amplifying factors (Sharma situational disempowerment)."
readme = "README.md"
requires-python = ">=3.10,<3.12"

dependencies = [
  "torch>=2.0.0",
  "transformers>=4.30.0",
  "peft>=0.5.0",
  "datasets>=2.0.0",
  "numpy>=1.24.0",
  "tqdm>=4.65.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "jaxtyping>=0.3.0,<=0.3.2",
  "beartype>=0.14.0",
  "scikit-learn>=1.3.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "huggingface_hub>=0.20.0",
  # Annotation pipeline
  "anthropic>=0.36.0",
  "simple-parsing>=0.1.5",
  "pydantic>=2.6.0",
  "rouge-score>=0.1.2",
  "safetytooling",
  # Demo
  "modal>=0.61.0",
  "plotly>=5.22.0",
  "streamlit>=1.33.0",
  "pandas>=2.1.0",
  "accelerate>=0.20.0",
  # Experiment tracking
  "wandb>=0.15.0",
  "sentence-transformers>=5.4.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = [ "." ]
include = [
  "auth_projection",
  "auth_projection.*",
  "utils",
  "utils.*",
  "hallucination_probes",
  "hallucination_probes.*",
]
namespaces = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
# uv will install safetytooling from this git repo when requested
safetytooling = { git = "https://github.com/safety-research/safety-tooling.git", rev = "main" }



